pyproject.toml
[project]
name = "agent-eval-phase-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "openai-agents>=0.14.1",
    "langchain>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "requests>=2.32.0",
    "tavily-python>=0.5.0",
]
